[feat] Enhance Trainer for Regression Tasks and Optimize Efficiency
- Updated Trainer to support training models for regression tasks.
- Improved efficiency by changing dtype to float16 and implementing mixed precision training.
- Modified Dataset to handle float16 dtype for better performance.

+++ spottheplace/ml/dataset.py
@@ -59,9 +59,9 @@ class ClassificationDataset(Dataset):
 
         if self.model_type == 'ViT':
             inputs = self.feature_extractor(images=image, return_tensors="pt")
-            image = inputs["pixel_values"].squeeze()
+            image = inputs["pixel_values"].squeeze().to(dtype=torch.float16)
         elif self.model_type == 'CNN':
-            image = self.transform(image)
+            image = self.transform(image).to(dtype=torch.float16)
         else:
             raise ValueError('Model type not supported')
 
@@ -92,7 +92,7 @@ class RegressionDataset(Dataset):
         self.df = df
         self.image_paths = self.df['image_path'].values
 
-        self.targets = torch.tensor(self.df[['long', 'lat']].values, dtype=torch.float32)
+        self.targets = torch.tensor(self.df[['long', 'lat']].values, dtype=torch.float16)
 
         self.feature_extractor = feature_extractor
         self.transform = transform
@@ -118,9 +118,9 @@ class RegressionDataset(Dataset):
 
         if self.model_type == 'ViT':
             inputs = self.feature_extractor(images=image, return_tensors="pt")
-            image = inputs["pixel_values"].squeeze()
+            image = inputs["pixel_values"].squeeze().to(dtype=torch.float16)
         elif self.model_type == 'CNN':
-            image = self.transform(image)
+            image = self.transform(image).to(dtype=torch.float16)
         else:
             raise ValueError('Model type not supported')
